# docs/swagger.json
{"openapi":"3.0.0","info":{"title":"OpenAPI for Library","version":"1.0.0","description":"Specification for OpenAPI based Library API","termsOfService":"https://github.com/murshidazher/terms/","contact":{"name":"Murshid Azher","url":"https://github.com/murshidazher","email":"hello@murshidazher.com"},"license":{"name":"MIT","url":"https://github.com/murshidazher/express-swagger-api/LICENSE"}},"servers":[{"url":"http://localhost:4000"}],"paths":{"/api/v1/books":{"get":{"summary":"Returns the list of all the books","tags":["Books"],"responses":{"200":{"description":"The list of the books","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Book"}}}}}}},"post":{"summary":"Create a new book","tags":["Books"],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Book"}}}},"responses":{"200":{"description":"The book was successfully created","headers":{"Location":{"description":"Location of the created book","schema":{"type":"string","format":"uri","example":"http://example.com/api/v1/books/{assinedIdValue}"}}},"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Book"}}}},"500":{"description":"Some server error"}}}},"/api/v1/books/{id}":{"get":{"summary":"Get the book by id","tags":["Books"],"parameters":[{"in":"path","name":"id","schema":{"type":"string"},"required":true,"description":"The book id"}],"responses":{"200":{"description":"The book description by id","contens":{"application/json":{"schema":{"$ref":"#/components/schemas/Book"}}}},"404":{"description":"The book was not found"}}},"put":{"summary":"Update the book by the id","tags":["Books"],"parameters":[{"in":"path","name":"id","schema":{"type":"string"},"required":true,"description":"The book id"}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Book"}}}},"responses":{"200":{"description":"The book was updated","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Book"}}}},"404":{"description":"The book was not found"},"500":{"description":"Some error happened"}}},"delete":{"summary":"Remove the book by id","tags":["Books"],"parameters":[{"in":"path","name":"id","schema":{"type":"string"},"required":true,"description":"The book id"}],"responses":{"200":{"description":"The book was deleted"},"404":{"description":"The book was not found"}}}},"/api/v1":{"get":{"summary":"Returns a response as pong","responses":{"200":{"description":"The initial heartbeat request","content":{"application/json":{"schema":{"type":"string","items":{"$ref":"#/components/schemas/Heartbeat"}}}}}}}}},"components":{"schemas":{"Book":{"type":"object","required":["title","author"],"properties":{"id":{"type":"string","format":"uuid","description":"The auto-generated id of the book","example":"d5fE_asz","readOnly":true},"title":{"type":"string","minLength":2,"maxLength":100,"example":"The New Turing Omnibus","description":"The book title"},"author":{"type":"string","minLength":2,"maxLength":100,"example":"Alexander K. Dewdney","description":"The book author"}}},"Heartbeat":{"type":"object","properties":{"response":{"type":"string","example":"pong","description":"The auto-generated response for heartbeat"}}}}},"tags":[{"name":"Books","description":"The books managing API"}]}